implement test_set_value
